+++ asm_f/srcs/parse_args.c
@@ -6,7 +6,7 @@
 /*   By: byermak <byermak@student.unit.ua>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/06/16 15:15:00 by byermak           #+#    #+#             */
-/*   Updated: 2018/06/16 15:15:25 by byermak          ###   ########.fr       */
+/*   Updated: 2018/07/02 16:21:09 by byermak          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -42,7 +42,8 @@ static int	parse_third_arg(char *str, t_arg **arg)
 	if ((i = skip_spaces(str)) == -1)
 		return (ERR_ENDLINE);
 	i = word(str, i);
-	if ((ret = (parse_arg(ft_strsub(str, g_x, i - g_x), arg))) != 1)
+	if ((ret = (parse_arg(ft_strsub(str,
+			(unsigned int)g_x, (size_t)(i - (int)g_x)), arg))) != 1)
 		return (ret);
 	g_x = (size_t)i;
 	return (1);
@@ -59,7 +60,8 @@ static int	parse_second_arg(char *str, t_arg **arg)
 	if ((i = skip_spaces(str)) == -1)
 		return (ERR_ENDLINE);
 	i = word(str, i);
-	if ((ret = (parse_arg(ft_strsub(str, g_x, i - g_x), arg))) != 1)
+	if ((ret = (parse_arg(ft_strsub(str,
+			(unsigned int)g_x, (size_t)(i - (int)g_x)), arg))) != 1)
 		return (ret);
 	g_x = (size_t)i;
 	return (1);
@@ -76,8 +78,8 @@ int			parse_args(char *str, t_code *new)
 	if ((i = skip_spaces(str)) == -1)
 		return (ERR_NO_COMMAND_ARGS);
 	i = word(str, i);
-	if ((ret = (parse_arg(ft_strsub(str, g_x, i - g_x), &(new->arg1)))) != 1 ||
-		(ret = check_first_arg(new)) != 1)
+	if ((ret = (parse_arg(ft_strsub(str, (unsigned int)g_x, (size_t)(i - (int)
+			g_x)), &(new->arg1)))) != 1 || (ret = check_first_arg(new)) != 1)
 		return (ret);
 	if ((g_x = (size_t)i) && (skip_spaces(str)) != -1)
 	{
